manage.py: Fix redundant prepare and multiple node procs spawning
The Django dev server starts seems to trigger a reload as soon as it
starts, which was causing prepare_server to get called twice.

Furthermore, a new node.js process was being spawned every time Django
reloaded.

+++ manage.py
@@ -22,8 +22,6 @@ except ImportError as exc:
 
 def prepare_server():
     """Apply migrations and collect static files."""
-    django.setup()
-
     print("Applying migrations.")
     call_command("migrate")
 
@@ -32,8 +30,10 @@ def prepare_server():
         call_command("collectstatic", interactive=False, clear=True)
 
 
-def run_dev_server():
-    """Simultaneously run the Django development server and ng build with --watch."""
+def run_ng_build():
+    """Run the Angular build in watch mode and return its process handle."""
+    print("Starting Angular build in watch mode.")
+
     node = shutil.which("node")
     if node is None:
         raise FileNotFoundError(
@@ -52,27 +52,45 @@ def run_dev_server():
             f"by running 'npm ci' while in {angular_root}"
         )
 
-    node_process = subprocess.Popen(
+    return subprocess.Popen(
         [node, str(ng), "build", "-c", "development", "--watch"],
         cwd=angular_root,
     )
+
+
+def run_dev_server():
+    """Simultaneously run the Django development server and ng build with --watch."""
+    # Avoid spawning a new node node process every time Django reloads due to a file change.
+    if os.environ.get("RUN_MAIN") != "true":
+        node_process = run_ng_build()
+    else:
+        node_process = None
 
     try:
         # TODO: terminate runserver if ng build crashes?
         call_command("runserver", "0.0.0.0:8000")
     except Exception:
         # If Django fails, terminate Angular's build too.
-        node_process.terminate()
+        if node_process:
+            node_process.terminate()
         raise
     else:
         # Wait for the process so that signals are forwarded if this wait is interrupted.
         # In practice, this may be unreachable since Django won't stop unless an exception happens.
-        node_process.wait()
+        if node_process:
+            node_process.wait()
 
 
 def run_server():
     """Prepare and run the web server."""
-    prepare_server()
+    django.setup()  # This must only be called once.
+
+    # `RUN_MAIN` is set to None the first time `runserver` is called. When the `use_reloader` option
+    # is set (it is set by default), `RUN_MAIN` is set to "true" upon a reload after a code change.
+    # To avoid preparing twice at the start, only prepare when in the reloader. Since `runserver`
+    # is only using during debug mode, the `RUN_MAIN` check only needs to happen in debug mode.
+    if not settings.DEBUG or os.environ.get("RUN_MAIN") == "true":
+        prepare_server()
 
     print("Starting server.")
